requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==1.3.5
textstat==0.7.3
nltk==3.8.1
